Fade fixes, title scrolling, model changes
Fixed issues with UI fading; excluded tooltips and popups. Frames now
have to be explicitly mouse enabled to fade back in.
Added custom quest progress content handler.
A quest marker is now shown instead of the player when a quest comes
from 'the ether'.
Buttons are now readjusted in height after scaling up/down to account
for UI scale inconsistency.
Buttons can now be scrolled vertically to account for the rare case when
they collectively exceed the height of the viewport.
The main content frame is now dynamically readjusting to whatever
content is displayed.